# internal/server/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "水杉输入法后端 API",
    "version": "1",
    "description": "水杉输入法共通后端。点击 Authorize 填写设备令牌（不含 Bearer 前缀）。功能是否启用请查询 capabilities。JSON 请求最多 64 KiB，multipart 总体最多 16 MiB。在线输入接口不保存输入和音频；用户同步接口按用户操作保存偏好及显式上传的数据。 用户接口详见用户体系标签；登录成功后也可使用用户 access_token 调用在线输入接口。"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "deviceToken": []
    },
    {
      "userSession": []
    }
  ],
  "paths": {
...
    "/v1/capabilities": {
      "get": {
        "operationId": "capabilities",
        "summary": "查询已启用能力",
        "tags": [
          "系统"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "api_version": {
                      "type": "string"
                    },
                    "cloud": {
                      "type": "boolean"
                    },
                    "chat": {
                      "type": "boolean"
                    },
                    "translation": {
                      "type": "boolean"
                    },
                    "transcription": {
                      "type": "boolean"
                    },
                    "streaming_transcription": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "api_version",
                    "cloud",
                    "chat",
                    "translation",
                    "transcription",
                    "streaming_transcription"
                  ]
                },
                "example": {
                  "api_version": "1",
                  "cloud": true,
                  "chat": true,
                  "translation": true,
                  "transcription": true,
                  "streaming_transcription": false
                }
              }
            }
          },
          "400": {
            "description": "输入参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "设备令牌缺失或无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "浏览器来源未获允许",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "音频上传超过大小限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "不支持的内容类型",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "设备请求触发限流；Retry-After 表示等待秒数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "重试等待秒数（繁忙时提供）",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "502": {
            "description": "上游请求失败；不暴露供应商响应正文",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "功能未启用或服务繁忙；繁忙时提供 Retry-After 等待秒数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "重试等待秒数（繁忙时提供）",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "504": {
            "description": "请求超过截止时间",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/translate": {
      "post": {
        "operationId": "translation",
        "summary": "候选翻译",
        "tags": [
          "在线输入"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "data"
                  ]
                },
                "example": {
                  "code": 200,
                  "data": "test"
                }
              }
            }
          },
          "400": {
            "description": "输入参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "设备令牌缺失或无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "浏览器来源未获允许",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "音频上传超过大小限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "不支持的内容类型",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "设备请求触发限流；Retry-After 表示等待秒数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "重试等待秒数（繁忙时提供）",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "502": {
            "description": "上游请求失败；不暴露供应商响应正文",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "功能未启用或服务繁忙；繁忙时提供 Retry-After 等待秒数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "重试等待秒数（繁忙时提供）",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "504": {
            "description": "请求超过截止时间",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "description": "非空文字，UTF-8 最多 8192 字节。"
                  },
                  "source_lang": {
                    "type": "string",
                    "pattern": "^[A-Za-z-]{2,16}$",
                    "example": "AUTO"
                  },
                  "target_lang": {
                    "type": "string",
                    "pattern": "^[A-Za-z-]{2,16}$",
                    "example": "EN"
                  }
                },
                "required": [
                  "text",
                  "source_lang",
                  "target_lang"
                ],
                "additionalProperties": false
              },
              "example": {
                "text": "测试",
                "source_lang": "AUTO",
                "target_lang": "EN"
              }
            }
          }
        }
      }
    },
    "/v1/audio/transcriptions": {
      "post": {
        "operationId": "transcription",
        "summary": "WAV 批量语音转写",
        "tags": [
          "在线输入"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "text": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "text"
                  ]
                },
                "example": {
                  "text": "测试"
                }
              }
            }
          },
          "400": {
            "description": "输入参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "设备令牌缺失或无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "浏览器来源未获允许",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "音频上传超过大小限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "不支持的内容类型",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "设备请求触发限流；Retry-After 表示等待秒数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "重试等待秒数（繁忙时提供）",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "502": {
            "description": "上游请求失败；不暴露供应商响应正文",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "功能未启用或服务繁忙；繁忙时提供 Retry-After 等待秒数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "重试等待秒数（繁忙时提供）",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "504": {
            "description": "请求超过截止时间",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "PCM/IEEE-float RIFF/WAVE，最多 15728640 字节。"
                  },
                  "model": {
                    "type": "string",
                    "description": "服务端覆盖此模型字段。"
                  },
                  "language": {
                    "type": "string",
                    "pattern": "^[A-Za-z-]{2,16}$",
                    "example": "zh"
                  },
                  "response_format": {
                    "type": "string",
                    "enum": [
                      "json"
                    ],
                    "default": "json"
                  }
                },
                "required": [
                  "file"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/v1/audio/stream": {
      "get": {
        "operationId": "streaming_transcription",
        "summary": "实时语音 WebSocket（仅文档）",
        "tags": [
          "实时语音"
        ],
        "description": "通过 WSS 传输豆包 ASR v1 二进制消息，保留消息边界和载荷；设备使用 Bearer 鉴权，供应商请求头由后端提供。\n\nSwagger UI 不支持 WebSocket 二进制会话。使用 WSS 客户端携带设备 Bearer 令牌建立连接。仅返回通用关闭原因，不透传上游 HTTP 正文或连接关闭说明。 单消息最多 1048576 字节，每方向每会话最多 33554432 字节。",
        "x-websocket": true,
        "responses": {
          "101": {
            "description": "WebSocket 升级成功；后续为豆包 ASR v1 二进制消息"
          },
          "400": {
            "description": "需要 WebSocket Upgrade"
          },
          "401": {
            "description": "设备令牌无效"
          },
          "503": {
            "description": "功能未启用或服务繁忙"
          }
        }
      }
    },
...
    "/v1/input/capabilities": {
      "get": {
        "summary": "查询公共引擎配置能力",
        "tags": [
          "公共输入引擎"
        ],
        "description": "返回 Engine 和词库是否配置，以及支持的输入方案、双拼方案和候选上限。",
        "responses": {
          "200": {
            "description": "成功"
          },
          "401": {
            "description": "缺少有效令牌"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "deviceToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "管理员发放的设备令牌；不是供应商密钥。"
      },
      "userSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "登录返回的 access_token，不是 refresh_token 或供应商密钥。"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}